Deprecation changes
I've explicitly made apt_remount_filesystems in to an empty list to solve a
"[DEPRECATION WARNING]: evaluating None as a bare variable" message, this
feature changes in 2.12.
While making that change I adjusted the description to be more explicit about
the filesystem options.

And at while doing that I've removed the backwards compatibility with
apt_remount_filesystem - it was meant to be short lived compat and due to
generating a deprecation warning of its own has now been removed.